Add placeholder step defs and implement earning table parsing

+++ src/SFA.DAS.Payments.AcceptanceTests/Refactoring/ReferenceDataModels/PeriodValue.cs
@@ -3,6 +3,6 @@
     public class PeriodValue
     {
         public string PeriodName { get; set; }
-        public long Value { get; set; }
+        public decimal Value { get; set; }
     }
 }
